+++ src/main/kotlin/com/github/strindberg/emacsj/actions/search/ISearchReplaceRegexpAction.kt
@@ -4,4 +4,6 @@ import com.github.strindberg.emacsj.search.ISearchReplaceHandler
 import com.github.strindberg.emacsj.search.SearchType
 import com.intellij.openapi.editor.actionSystem.EditorAction
 
-class ISearchReplaceRegexpAction : EditorAction(ISearchReplaceHandler(SearchType.REGEXP)), ISearchAction
+class ISearchReplaceRegexpAction :
+    EditorAction(ISearchReplaceHandler(SearchType.REGEXP)),
+    ISearchAction