+++ src/main/kotlin/com/github/strindberg/emacsj/actions/search/ISearchReplaceTextAction.kt
@@ -4,4 +4,6 @@ import com.github.strindberg.emacsj.search.ISearchReplaceHandler
 import com.github.strindberg.emacsj.search.SearchType
 import com.intellij.openapi.editor.actionSystem.EditorAction
 
-class ISearchReplaceTextAction : EditorAction(ISearchReplaceHandler(SearchType.TEXT)), ISearchAction
+class ISearchReplaceTextAction :
+    EditorAction(ISearchReplaceHandler(SearchType.TEXT)),
+    ISearchAction